requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
drf-spectacular==0.25.1
drf-spectacular-sidecar==2022.12.1